Prevent teachers from grading students by accident.
To unlock submissions, teachers need to visit the grading page. If
they clicked "Save changes", they would assign a grade of zero to
the group. This patch requires a grade to be given for grades to
be saved. If a grade is not provided, some fields are hidden and
the teacher will be redirected to the main page.

When a grade is provided, and changes are saved, a message will
be displayed to the teacher informing them that the grades were saved.
At this point, teachers can no longer remove the grade when submitting
the form.

Additionally, the unlocking confirmation message was made clearer
that unlocking would take effect immediately.