Update calculator.php
The old algorithm used a "fudge factor" to deal with complete non-submissions. This unfairly favors non-submitters since it always increases their PA marks. (They receive all the same marks from peers, but now multiplied by a fudge factor that is greater than one.) This fix sets the fudge factor to one, and replaces blank submissions with straight ones. This solves the problem while not influencing either the non-submitter's marks, or the marks of their peers.